requirements.txt
# Core ML/DL Libraries
torch>=2.0.0
transformers>=4.30.0
scipy>=1.10.0
scikit-learn>=1.3.0
numpy>=1.24.0
pandas>=2.0.0
imbalanced-learn>=0.11.0  # For SMOTE class balancing

# Graph/Tree Processing
networkx>=3.0
torch-geometric>=2.3.0

# Utilities
tqdm>=4.65.0
PyYAML>=6.0
matplotlib>=3.7.0
seaborn>=0.12.0
tensorboard>=2.13.0

# Optimization
accelerate>=0.20.0
datasets

# Embedding Models (Qodo-Embed-1-1.5B)
sentence-transformers>=2.2.2
